Bring update summary options to all update-related goals of the maven plugin
 #4395 and #5077 brought the options `showSummary` and `showSummaryOutput` to the `update` goal of the maven plugin.

These options seem to be relevant for the `updateSQL` and `updateTestingRollback` goals as well, so this PR brings them to the parent abstract class of all these goals
