+++ liquibase-maven-plugin/src/main/java/org/liquibase/maven/plugins/LiquibaseUpdate.java
@@ -1,9 +1,7 @@
 package org.liquibase.maven.plugins;
 
 import liquibase.*;
-import liquibase.database.Database;
 import liquibase.exception.LiquibaseException;
-import org.apache.maven.plugin.MojoExecutionException;
 import org.liquibase.maven.property.PropertyElement;
 
 /**
@@ -23,24 +21,6 @@ public class LiquibaseUpdate extends AbstractLiquibaseUpdateMojo {
      */
     @PropertyElement
     protected boolean dropFirst;
-
-    /**
-     * Whether or not to print a summary of the update operation.
-     * Allowed values: 'OFF', 'SUMMARY' (default), 'VERBOSE'
-     *
-     * @parameter property="liquibase.showSummary"
-     */
-    @PropertyElement
-    protected UpdateSummaryEnum showSummary;
-
-    /**
-     * Flag to control where we show the summary.
-     * Allowed values: 'LOG', 'CONSOLE', OR 'ALL' (default)
-     *
-     * @parameter property="liquibase.showSummaryOutput"
-     */
-    @PropertyElement
-    protected UpdateSummaryOutputEnum showSummaryOutput;
 
     @Override
     protected void doUpdate(Liquibase liquibase) throws LiquibaseException {
@@ -63,14 +43,6 @@ public class LiquibaseUpdate extends AbstractLiquibaseUpdateMojo {
                 throw new LiquibaseException(exception);
             }
         }
-    }
-
-    @Override
-    protected Liquibase createLiquibase(Database db) throws MojoExecutionException {
-        Liquibase liquibase = super.createLiquibase(db);
-        liquibase.setShowSummary(showSummary);
-        liquibase.setShowSummaryOutput(showSummaryOutput);
-        return liquibase;
     }
 
     @Override
